test(ir): add list folding IR tests
